Fixed none type issue and normal flow executes
