+++ SupplyHero_FastApi/features/steps/ID003_steps.py
@@ -1,7 +1,7 @@
 from behave import *
 import jwt
 import fastapi_users
-from db.mongo_regular import MongoSessionRegular
+from db.mongo import MongoSessionRegular
 from fastapi.testclient import TestClient
 from api import main
 from api.config import SECRET
@@ -18,7 +18,7 @@ def step_impl(context, email, password):
     context.register = {"email": email, "password": password}
     context.test_client = TestClient(main.app)
     context.response = context.test_client.post("/register", json=context.register)
-    print(context.response.json())
+    # print(context.response.json())
     context.user = email
     context.password = password
 
@@ -27,7 +27,7 @@ def step_impl(context, email, password):
 def step_impl(context, email, password):
     context.login = {"username": email, "password": password}
     context.response = context.test_client.post("/login", data=context.login)
-    print(context.response.json())
+    # print(context.response.json())
     context.access_token = context.response.json()["access_token"]
 
 
@@ -37,16 +37,15 @@ def step_impl(context, email, password):
     uuid_object = UUID(context.data["user_id"])
     user = context.mongo_sesh.find_json({"id": uuid_object})
     found_user = user["email"]
-    found_pass = fastapi_users.password.get_password_hash(context.password)
-    print(user)
-    print(found_pass)
+    found_hashed_pass = user['hashed_password']
+    verified, __ = fastapi_users.password.verify_and_update_password(password, found_hashed_pass)
     assert found_user == email
-    assert found_pass == password
+    assert verified
 
 
 @then('the user has logged in successfully')
 def step_impl(context):
     context.status_code = context.response.status_code
     #TODO: need cleanup to remove entries from db
-    context.mongo.sesh.delete_json({"user_id": context.data["user_id"]})
+    context.mongo_sesh.delete_json({"user_id": context.data["user_id"]})
     assert context.status_code == 200
